Remove python 2/3 differences in code
There previously were only two more methods six didn't provide and we
did: b2s and u2s. These were defined separately for pyton 2.7 and 3,
which isn't necessary anymore. Moreover the u2s method is completely
unnecessary now, and since it was used only on a single module that
module was updated as well.

The b2s method could also still be removed in theory, but it would
require touching quite a few more lines of code in many modules, so I'll
refrain from doing it at the moment to keep changes small.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>

Former-commit-id: 876d962008067706a8f07faa4d61d840a6e4c9db
Former-commit-id: 2db566aaa34d4ac7b207101b509146d784f0a816