Bring back threadpool-based parallel drop execution
The current implementation of multi-process drop execution was flawed
because it created a process per drop instance, leaving the fate of the
underlying system to the mercy of the size of the graph being executed.
Because of this, RPC clients were also created for every remote drop
instead of for every single remote host, which again can exhaust
resources in the host fairly rapidly.

This commit removes all traces of multi-process drop execution, bringing
back the simpler-but-reliable thread-based drop execution. By default
drops create their own threads when wanting to async_execute, although
differently from before this now happens via a WorkerPool class that
takes care of doing this. The NodeManager overrides that behavior by
injecting its own WorkerPool that is *always* backed by a thread pool
(which wasn't the case before, we left the thread count go wild).

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>